[move-prover] use the new TypeUnifier scheme to simulate the old behavior
If a global invariant `I` talks about a generic resource, e.g., `S<X, Y>`,
and a function `f` might also be parameterized `f<X, Y>`. Then, there
should be a procedure to decide whether a global invariant `I` should
be checked in the body of `f`.

To put into the terms of `TypeUnifier`, we have the following relation
in the `invariants_with_generics.move` test case:

```
                        F_inst                  I_inst
f<u64, bool>
  I<u64, bool>          -                       -
  I<u64, Y2>            -                       Y2 := bool
  I<Y1, bool>           -                       Y1 := u64
  I<Y1, Y2>             -                       Y1 := u64, Y2 := bool

f<u64, X2>
  I<u64, bool>          X2 := bool              -
  I<u64, Y2>            -                       -
  I<Y1, bool>           X2 := bool              Y1 := u64
  I<Y1, Y2>             -                       Y1 := u64

f<X1, bool>
  I<u64, bool>          X1 := u64               -
  I<u64, Y2>            X1 := u64               Y2 := bool
  I<Y1, bool>           -                       -
  I<Y1, Y2>             -                       Y2 := bool

f<X1, X2>
  I<u64, bool>          X1 := u64, X2 := bool   -
  I<u64, Y2>            X1 := U64               -
  I<Y1, bool>           X2 := bool              -
  I<Y1, Y2>             -                       -
```

In the old behavior (and in current simulation), only the ones marked
with `-` on the `F_inst` side are checked (i.e., I is checked in f)
but in theory, all pairs should be checked.